Module wireguard_builder_rs::models

The GraphQL schema

Re-exports

pub use client::Client;

pub use server::Server;

pub use user::JwtUser;

pub use user::User;

Modules

client

Module that holds everything that is necessary for the Client

dns_server

Module that holds everything that is necessary for the DnsServer

keypair

Module that holds everything that is necessary for the Keypair type

server

Module that holds everything that is necessary for the Server

user

Holds all the necessary things for a User

vpn_ip_address

Module that holds everything that is necessary for the VpnIpAddress

vpn_network

Module that holds everything that is necessary for the VpnNetwork

Structs

Mutation

The root of the mutation type

QueryRoot

The root of the Query type

Functions

create_connection

Retrieves a single database connection from the database connection pool and returns it

create_schema

Creates a new schema with a connection pool for communicating with the database as context

Type Definitions

GrahpQLSchema

Represents the schema that is created by create_schema()