Struct wireguard_builder_rs::models::client::Client

pub struct Client {
    id: i32,
    name: String,
    description: Option<String>,
    keepalive_interval: i32,
}

Fields
id: i32name: Stringdescription: Option<String>keepalive_interval: i32

The interval in seconds where the client should reconnect to the server

Implementations

impl Client

async fn id(&self, ctx: &Context<'_>) -> Result<&i32>

async fn name(&self, ctx: &Context<'_>) -> Result<&String>

async fn description(&self, ctx: &Context<'_>) -> Result<&Option<String>>

async fn keepalive_interval(&self, ctx: &Context<'_>) -> Result<&i32>

impl Client

async fn config(&self, ctx: &Context<'_>) -> Result<Option<String>>

A wireguard configuration for the Client

async fn keypair(&self, ctx: &Context<'_>) -> Result<Keypair>

The keypair that is used by the client

async fn dns_server(&self, ctx: &Context<'_>) -> Result<DnsServer>

The dns server that is used by the client

async fn vpn_network(&self, ctx: &Context<'_>) -> Result<VpnNetwork>

The vpn network that the client is associated with

async fn ip_address(&self, ctx: &Context<'_>) -> Result<String>

The ip address of the client in the vpn network

async fn server(&self, ctx: &Context<'_>) -> Result<Option<Server>>

impl Client

pub fn delete(
connection: &PooledConnection<ConnectionManager<PgConnection>>,
id: i32
) -> Result<bool>

Deletes the client with the given id from the database

Arguments

connection - A connection to the database

id - The id of the client that should be deleted

pub fn create(
connection: &PooledConnection<ConnectionManager<PgConnection>>,
client: &InputClient
) -> Result<QueryableClient>

Creates a new client in the database

Arguments

connection - A connection to the database

client - The client that should be inserted into the database

Returns

Returns [Result::Ok] if the operation was a success. If validation of the input parameters fails an [Result::Error] is returned.

pub fn get_keypair_ids(
connection: &PooledConnection<ConnectionManager<PgConnection>>
) -> Result<Vec<i32>>

Returns all the ids of keypairs that are used by a client

pub fn get_by_id(
connection: &PooledConnection<ConnectionManager<PgConnection>>,
client_id: i32
) -> Result<QueryableClient>

Returns the client for the given id

Arguments

connection - A connection to the database

client_id - The id of the client that should be queried

Panics

Panics if no client was found