Struct wireguard_builder_rs::models::keypair::Keypair

pub struct Keypair {
    pub id: i32,
    pub public_key: String,
    pub private_key: String,
}

A Keypair that is used by a Client or Server

Fields
id: i32public_key: Stringprivate_key: String
Implementations

impl Keypair

pub async fn id(&self, ctx: &Context<'_>) -> Result<&i32>

pub async fn public_key(&self, ctx: &Context<'_>) -> Result<&String>

pub async fn private_key(&self, ctx: &Context<'_>) -> Result<&String>

impl Keypair

pub fn create<'a>(
connection: &PooledConnection<ConnectionManager<PgConnection>>,
public_key: &'a str,
private_key: &'a str
) -> Keypair

pub fn delete(
connection: &PooledConnection<ConnectionManager<PgConnection>>,
keypair_id: i32
) -> Result<bool>

Deletes a Keypair from the database

Arguments

connection - A connection to the database

keypair_id - The id of the Keypair that should be deleted

Returns

Returns true if the operation was successful or an error

pub fn generate_keypair() -> (String, String)

Generates a new Keypair and returns it

pub fn get_by_id(
connection: &PooledConnection<ConnectionManager<PgConnection>>,
keypair_id: i32
) -> Result<Keypair>

Returns the Keypair for the given id

Arguments

connection - A connection to the database

keypair_id - The id of the Keypair that should be returned

Returns

Returns the Keypair or an Result::Error if the Keypair does not exist or the query failed