Struct wireguard_builder_rs::models::vpn_ip_address::VpnIpAddress

pub struct VpnIpAddress {
    pub id: i32,
    pub vpn_network_id: i32,
    pub ip_address: String,
}

A VpnIpAddress represents a unique ip address that is associated with a VpnNetwork

Fields
id: i32vpn_network_id: i32ip_address: String
Implementations

impl VpnIpAddress

pub fn create<'a>(
connection: &PooledConnection<ConnectionManager<PgConnection>>,
vpn_network_id: i32,
ip_address: &'a str
) -> Result<VpnIpAddress>

Creates a new VpnIpAddress

Arguments

connection - A connection to the database

vpn_network_id - The id of the VpnNetwork with that the ip address should be associated with

ip_address - The ip address that should be created

Returns

Returns the created VpnIpAddress or an error from the database

pub fn delete<'a>(
connection: &PooledConnection<ConnectionManager<PgConnection>>,
vpn_ip_address_id: i32
) -> Result<bool>

Deletes a VpnIpAddress from the database

Arguments

connection - A connection to the database

vpn_ip_address_id - The id of the VpnIpAddress that should be deleted

Returns

Returns true if the operation was successful or an error

pub fn get_by_id(
connection: &PooledConnection<ConnectionManager<PgConnection>>,
ip_id: i32
) -> VpnIpAddress

Returns the VpnIpAddress for the given id

Arguments

connection - A connection to the database

ip_id - The id of the VpnIpAddress that should be returned

Panics

Panics if no ip address was found