Struct wireguard_builder_rs::models::vpn_network::VpnNetwork

pub struct VpnNetwork {
    pub id: i32,
    pub name: String,
    pub description: Option<String>,
    pub ip_network: String,
    pub subnetmask: i32,
    pub listen_port: i32,
    pub interface_name: String,
}

A VpnNetwork represents a network that contains clients and a server

Fields
id: i32name: Stringdescription: Option<String>ip_network: String

The ip_network that the object represents

subnetmask: i32

The subnetmask of the ip_network in CIDR format

listen_port: i32

The port where the vpn server is listening on

interface_name: String

The name of the interface (e.g. wg0)

Implementations

impl VpnNetwork

pub async fn id(&self, ctx: &Context<'_>) -> Result<&i32>

pub async fn name(&self, ctx: &Context<'_>) -> Result<&String>

pub async fn description(&self, ctx: &Context<'_>) -> Result<&Option<String>>

pub async fn ip_network(&self, ctx: &Context<'_>) -> Result<&String>

pub async fn subnetmask(&self, ctx: &Context<'_>) -> Result<&i32>

pub async fn listen_port(&self, ctx: &Context<'_>) -> Result<&i32>

pub async fn interface_name(&self, ctx: &Context<'_>) -> Result<&String>

impl VpnNetwork

async fn clients(&self, ctx: &Context<'_>) -> Result<Vec<Client>>

All Clients that are associated with the VpnNetwork

async fn server(&self, ctx: &Context<'_>) -> Result<Option<Server>>

The Server that is associated with the VpnNetwork

impl VpnNetwork

pub fn create(
connection: &PooledConnection<ConnectionManager<PgConnection>>,
vpn_network: &InputVpnNetwork
) -> Result<VpnNetwork>

Creates a new VpnNetwork in the database

Arguments

connection - A connection to the database

vpn_network - The VpnNetwork that should be created

Returns

Returns the created VpnNetwork or an error that was thrown by the database

pub fn update(
connection: &PooledConnection<ConnectionManager<PgConnection>>,
net_id: i32,
vpn_network: &InputVpnNetwork
) -> Result<VpnNetwork>

Updates a VpnNetwork in the database

Arguments

connection - A connection to the database

net_id - The id of the VpnNetwork that should be updated

vpn_network - The new VpnNetwork object that should replace the old one

Returns

The update may return an error if the new values violate unique constraints in the database. Otherwise the updated vpn network is returned.

pub fn delete(
connection: &PooledConnection<ConnectionManager<PgConnection>>,
net_id: i32
) -> Result<bool>

Deletes a VpnNetwork from the database

Arguments

connection - A connection to the database

net_id - The id of the VpnNetwork that should be deleted

Returns

Returns true if the delete action was successful an error otherwise

pub fn get_by_id(
connection: &PooledConnection<ConnectionManager<PgConnection>>,
net_id: i32
) -> Option<VpnNetwork>

Returns the VpnNetwork for the given id

Arguments

connection - A connection to the database

net_id - The id of the vpn network that should be returned

Returns

If a vpn network was found a Option::Some will be returned Option::None otherwise

pub fn get_associated_clients(
&self,
connection: &PooledConnection<ConnectionManager<PgConnection>>
) -> Option<Vec<QueryableClient>>

Returns the Clients that are associated with the VpnNetwork

pub fn get_associated_server(
&self,
connection: &PooledConnection<ConnectionManager<PgConnection>>
) -> Option<QueryableServer>

Returns the Server that is associated with the VpnNetwork