Module wireguard_builder_rs::schema::clients::dsl

Re-exports all of the columns of this table, as well as the table struct renamed to the module name. This is meant to be glob imported for functions which only deal with one table.

Re-exports

pub use super::columns::id;

pub use super::columns::name;

pub use super::columns::description;

pub use super::columns::keepalive_interval;

pub use super::columns::dns_server_id;

pub use super::columns::keypair_id;

pub use super::columns::vpn_ip_address_id;

pub use super::table as clients;

Macros

__static_cond

__static_cond

__static_cond

__static_cond

__static_cond

__static_cond

__static_cond