Module wireguard_builder_rs::schemas

The GraphQL schema

Modules

dns_server

Module that holds everything that is necessary for the DnsServer

keypair

Module that holds everything that is necessary for the Keypairs

vpn_network

Module that holds everything that is necessary for the VpnNetwork

Structs

Mutation

The root of the mutation type

QueryRoot

The root of the Query type

Functions

create_connection

Retrieves a single database connection from the database connection pool and returns it

create_schema

Creates a new schema with a connection pool for communicating with the database as context

Type Definitions

DatabaseConnection

Represents a pool of connections to the database

GrahpQLSchema

Represents the schema that is created by create_schema()

SingleConnection

Represents a single connection to the database