Function wireguard_builder_rs::validate::is_ip_in_network

pub fn is_ip_in_network(
    ip_network: Ipv4Addr, 
    subnetmask: i32, 
    ip_address: Ipv4Addr
) -> bool

Validates if an ip address is in the given network or not

Arguments

ip_network - The ip network in that the ip addresss should be in

subnetmask - The subnetmask of the ip_network in CIDR format

ip_address - The ip address that should be validated